Try a different split of pretty text field and helper  * This one puts everything into the helper, and leaves the field as a simple wrapper
