+++ lib/components/helpers/pretty-text-helper.js
@@ -48,17 +48,19 @@ module.exports = React.createClass({
 
     return {
       value: this.props.value,
-      codeMirrorMode: this.props.isEditable,
+      codeMirrorMode: false,
+      isChoicesOpen: false,
       replaceChoices: replaceChoices,
       translator: translator
     };
   },
 
   componentWillReceiveProps: function(nextProps) {
+    var selectedChoices = this.props.selectedChoices;
+    var replaceChoices = this.props.replaceChoices;
     var nextState = {
-      replaceChoices: nextProps.replaceChoices,
-      translator: TagTranslator(this.props.selectedChoices.concat(nextProps.replaceChoices), this.props.config.humanize),
-      codeMirrorMode: nextProps.isEditable
+      replaceChoices: replaceChoices,
+      translator: TagTranslator(selectedChoices.concat(replaceChoices), this.props.config.humanize)
     };
 
     if (this.state.value !== nextProps.value && nextProps.value) {
@@ -68,20 +70,77 @@ module.exports = React.createClass({
     this.setState(nextState);
   },
 
+  handleChoiceSelection: function (key) {
+    var pos = this.state.selectedTagPos;
+    var tag = '{{' + key + '}}';
+
+    if (pos) {
+      this.codeMirror.replaceRange(tag, {line: pos.line, ch: pos.start}, {line: pos.line, ch: pos.stop});
+    } else {
+      this.codeMirror.replaceSelection(tag, 'end');
+    }
+    this.codeMirror.focus();
+
+    this.setState({ isChoicesOpen: false, selectedTagPos: null });
+  },
+
   render: function() {
     return this.renderWithConfig();
   },
 
   renderDefault: function () {
+
     var config = this.props.config;
     var textBoxClasses = cx(_.extend({}, this.props.classes, {'pretty-text-box': true}));
 
+    var onInsertClick = function () {
+      this.setState({selectedTagPos: null});
+      this.onToggleChoices();
+    };
+    var insertBtn = config.createElement('insert-button',
+                                         {ref: 'toggle', onClick: onInsertClick.bind(this)},
+                                         'Insert...');
+
+    var choices = config.createElement('choices', {
+      ref: 'choices',
+      choices: this.state.replaceChoices,
+      open: this.state.isChoicesOpen,
+      ignoreCloseNodes: this.getCloseIgnoreNodes,
+      onSelect: this.handleChoiceSelection,
+      onClose: this.onCloseChoices,
+      isAccordion: this.props.isAccordion
+    });
+
     // Render read-only version.
     return (
-      <div className={textBoxClasses} tabIndex={this.props.tabIndex} onFocus={this.props.onFocus} onBlur={this.props.onBlur}>
-        <div ref='textBox' className='internal-text-wrapper' />
+      <div className={cx({'pretty-text-wrapper': true, 'choices-open': this.state.isChoicesOpen})} onMouseEnter={this.switchToCodeMirror}>
+        <div className={textBoxClasses} tabIndex={this.props.tabIndex} onFocus={this.props.onFocus} onBlur={this.props.onBlur}>
+          <div ref='textBox' className='internal-text-wrapper' />
+        </div>
+        {insertBtn}
+        {choices}
       </div>
     );
+  },
+
+  getCloseIgnoreNodes: function () {
+    return this.refs.toggle.getDOMNode();
+  },
+
+  onToggleChoices: function () {
+    this.setChoicesOpen(!this.state.isChoicesOpen);
+  },
+
+  setChoicesOpen: function (isOpen) {
+    var action = isOpen ? 'open-replacements' : 'close-replacements';
+    this.props.onStartAction(action);
+    this.setState({ isChoicesOpen: isOpen });
+  },
+
+  onCloseChoices: function () {
+    if (this.state.isChoicesOpen) {
+      this.setChoicesOpen(false);
+    }
   },
 
   createEditor: function () {
@@ -141,15 +200,6 @@ module.exports = React.createClass({
     this.tagCodeMirror();
   },
 
-  insertText: function(text, pos) {
-    if (pos) {
-      this.codeMirror.replaceRange(text, {line: pos.line, ch: pos.start}, {line: pos.line, ch: pos.stop});
-    } else {
-      this.codeMirror.replaceSelection(text, 'end');
-    }
-    this.codeMirror.focus();
-  },
-
   createReadonlyEditor: function () {
     var textBoxNode = this.refs.textBox.getDOMNode();
 
@@ -174,13 +224,20 @@ module.exports = React.createClass({
     this.codeMirror = null;
   },
 
+  switchToCodeMirror: function () {
+    if (!this.state.codeMirrorMode) {
+      this.setState({codeMirrorMode: true});
+    }
+  },
+
   createTagNode: function (pos) {
     var node = document.createElement('span');
     var label = this.state.translator.getLabel(pos.tag);
     var config = this.props.config;
 
     var onTagClick = function () {
-      this.props.onTagClick(pos);
+      this.setState({selectedTagPos: pos});
+      this.onToggleChoices();
     };
 
     var props = {tag: pos.tag, replaceChoices: this.state.replaceChoices, onClick: onTagClick.bind(this)};
